# pkg/response/responses.go
package response

type SaveProfileResponse struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

type ErrorResponse struct {
	Error string `json:"error" validate:"required"`
}